Assignments 01 to 06/Project_09_Build a Python Website in 15 Minutes With Streamlit/requirements.txt
streamlit
Pillow
requests
# If you want to use the image upload feature, you may need to install additional libraries